fix: inconsistent @types/node reference
In this commit the dependency of @types/node is added to the dependencies with a `*` reference as is suggested by DefinitelyTyped. We also used 10.17.9 in mycloud and this commit uses this as well. Furthermore it addresses type differences in various node versions. It should be compatible even if we upgrade @types/node in mycloud in future.